Create an MNB-styled GT Table

This function creates a GT table with MNB (Magyar Nemzeti Bank) styling applied. It formats numbers, percentages, and applies specific styles to the table.

Usage

gt_mnb(data, title = NULL, comment = NULL)

Arguments

data

A data frame or GT table object to be styled.

title

Optional. A string to be used as the table title.

comment

Optional. A string to be added as a source note to the table.

Value

A GT table object with MNB styling applied.

Examples

iris |> 
  head() |> 
  gt_mnb(title = "Random title", comment = "Source: This is a built-in dataset")

  

Random title
    

Sepal.length
      	Sepal.width
      	Petal.length
      	Petal.width
      	Species
    

5,10
	3,50
	1,40
	20,00%
	setosa

4,90
	3,00
	1,40
	20,00%
	setosa

4,70
	3,20
	1,30
	20,00%
	setosa

4,60
	3,10
	1,50
	20,00%
	setosa

5,00
	3,60
	1,40
	20,00%
	setosa

5,40
	3,90
	1,70
	40,00%
	setosa

Source: This is a built-in dataset